move e2e-env to tests/e2e-env
