+++ tests/e2e-env/.npmrc
