+++ tests/e2e-env/bin/install-wp-tests.sh
