+++ tests/e2e-env/docker/wp-cli/entrypoint.sh
