+++ tests/e2e-env/docker/wp-cli/wait-for-it.sh
