+++ tests/e2e-env/package.json
@@ -10,13 +10,13 @@
     "e2e",
     "puppeteer"
   ],
-  "homepage": "https://github.com/woocommerce/woocommerce-admin/tree/master/packages/e2e-env/README.md",
+  "homepage": "https://github.com/woocommerce/woocommerce/tree/master/tests/e2e-env/README.md",
   "repository": {
     "type": "git",
-    "url": "https://github.com/woocommerce/woocommerce-admin.git"
+    "url": "https://github.com/woocommerce/woocommerce.git"
   },
   "bugs": {
-    "url": "https://github.com/woocommerce/woocommerce-admin/issues"
+    "url": "https://github.com/woocommerce/woocommerce/issues"
   },
   "main": "index.js",
   "devDependencies": {
