+++ webpack.config.js
@@ -1,0 +1,9 @@
+const { webpackAlias: coreE2EAlias } = require( '@woocommerce/e2e-env' );
+
+module.exports = {
+	resolve: {
+		alias: {
+			...coreE2EAlias,
+		},
+	},
+};
